+++ spring-mail/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.0
+
+- **breaking**: refactor dependency inject ([#105])
+
+[#105]: https://github.com/spring-rs/spring-rs/pull/105
+
 ## 0.2.2
 
 - **changed**: fix smtp tls transport bug ([#103])